refactor(raijin): remove exception filename redundancy
