+++ yarn/raijin/src/yarn/command.ts
@@ -2,7 +2,7 @@ import type { YarnCommandRunner }     from './interfaces.js'
 
 import { spawn }                      from 'node:child_process'
 
-import { RaijinYarnCommandException } from './exceptions/command.exception.js'
+import { RaijinYarnCommandException } from './exceptions/command.js'
 
 export const createYarnCommandEnvironment = (
   environment: NodeJS.ProcessEnv = process.env
