+++ yarn/raijin/src/yarn/exceptions/command.ts
